==> ZhouZhiwei/kansei_bj_sh_hotel/code/1_data_process.ipynb <==
"""
数据处理
"""
import pandas as pd

data = pd.read_excel('../data/Beijing & Shanghai - 4&5 star.xlsx')
data

data = data[~ data['评论文本'].isnull()]
data.shape

import emoji

data['评论文本'] = data['评论文本'].apply(lambda x: "".join(emoji.replace_emoji(x).split()))
data

data = data[~ data['评论文本'].str.contains(r'[a-zA-z]+://[^\s]*')]
data.shape

data = data[~ data['评论文本'].isna()]
data.shape
