==> ZhouZhiwei/kansei_bj_sh_hotel/code/2_LAC_pos.ipynb <==
"""
POS标注，使用百度LAC库
"""
import pandas as pd
from LAC import LAC

data = pd.read_excel('../data/data_cleaned.xlsx')
data

# POS报错：TypeError: object of type 'float' has no len()
# 数据中存在NaN数据
# data[~ data['评论文本'].apply(lambda x: isinstance(x, str))]
data = data[~ data['评论文本'].isna()]
data.shape

# baidu/LAC词性标注
lac = LAC(mode='lac')
data['LAC'] = data['评论文本'].apply(lambda x: lac.run(x))
data

data['word'], data['part'] = data['LAC'].apply(lambda x: x[0]), data['LAC'].apply(lambda x: x[1])
data

# 去停用词
with open('../data/hit_stopwords.txt', 'r', encoding='utf-8') as f:
    stop_words = f.readlines()
stop_words = [stop_word.strip('\n') for stop_word in stop_words]
data['stoped_word'] = data['word'].apply(lambda x: list(filter(lambda word: word not in stop_words, x)))
data

def get_nouns(lac):
    word, part = lac[0], lac[1]
    nouns = []
    for word_token, part_token in zip(word, part):
        if part_token == 'n':
            nouns.append(word_token)
    return nouns


data['nouns'] = data['LAC'].apply(get_nouns)
data

data['word2vec_corpus'] = data['stoped_word'].apply(lambda x: ' '.join(x))
data

data['nouns_corpus'] = data['nouns'].apply(lambda x: ' '.join(x))
data

data.to_excel('../data/data_POS_corpora.xlsx', index=False)

data['word2vec_corpus'].to_csv('../data/word2vec_corpus.txt', index=False)

data['nouns_corpus'].to_csv('../data/nouns.txt', index=False, header=False)

data['word'].apply(lambda x: ' '.join(x)).to_csv('../data/word2vec_corpus_with_stoped_punctuation.txt', index=False, header=False)
