==> ZhouZhiwei/kansei_bj_sh_hotel/code/6_word2vec_min_count_100.ipynb <==
"""
生成word2vec词向量，min_count=100
"""
import logging

logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

from gensim import utils


class MyCorpus:
    def __iter__(self):
        corpus_path = '../data/word2vec_corpus_with_stoped_punctuation.txt'
        for line in open(corpus_path, encoding='utf-8'):
            yield utils.simple_preprocess(line)

import gensim.models

sentences = MyCorpus()
# CBOW 200
cbow200 = gensim.models.Word2Vec(sentences=sentences, vector_size=200, min_count=100)

import tempfile

with tempfile.NamedTemporaryFile(prefix='gensim-model-', delete=False) as tmp:
    temporary_filepath = tmp.name
    cbow200.save(temporary_filepath)

# CBOW 300
cbow300 = gensim.models.Word2Vec(sentences=sentences, vector_size=300, min_count=100)

with tempfile.NamedTemporaryFile(prefix='gensim-model-', delete=False) as tmp:
    temporary_filepath = tmp.name
    cbow300.save(temporary_filepath)

# SG 200
sg200 = gensim.models.Word2Vec(sentences=sentences, vector_size=200, sg=1, min_count=100)

with tempfile.NamedTemporaryFile(prefix='gensim-model-', delete=False) as tmp:
    temporary_filepath = tmp.name
    sg200.save(temporary_filepath)

# SG 300
sg300 = gensim.models.Word2Vec(sentences=sentences, vector_size=300, sg=1, min_count=100)

with tempfile.NamedTemporaryFile(prefix='gensim-model-', delete=False) as tmp:
    temporary_filepath = tmp.name
    sg300.save(temporary_filepath)

# 计算不同模型的名词词向量：CBOW200, CBOW300 SG200 SG300
import pandas as pd

nouns = pd.read_csv('../data/nouns.txt')
nouns

nouns.columns = ['nouns']
nouns

nouns = nouns[~ nouns['nouns'].isna()]
nouns

# 拆分函数
import numpy as np


def split_row(data, column):
    """拆分成行

    :param data: 原始数据
    :param column: 拆分的列名
    :type data: pandas.core.frame.DataFrame
    :type column: str
    """
    row_len = list(map(len, data[column].values))
    rows = []
    for i in data.columns:
        if i == column:
            row = np.concatenate(data[i].values)
        else:
            row = np.repeat(data[i].values, row_len)
        rows.append(row)
    return pd.DataFrame(np.dstack(tuple(rows))[0], columns=data.columns)

nouns['temp'] = '1'
nouns

nouns_split = nouns.drop("nouns", axis=1).join(
    nouns["nouns"].str.split(" ", expand=True).stack().reset_index(level=1, drop=True).rename("noun"))
nouns_split

nouns_split = nouns_split.drop('temp', axis=1)
nouns_split

nouns_split = nouns_split.drop_duplicates(subset='noun')
nouns_split

nouns_split = nouns_split[nouns_split['noun'].apply(lambda x: x in cbow200.wv.index_to_key)]
nouns_split

nouns_split['CBOW200'], nouns_split['CBOW300'], nouns_split['SG200'], nouns_split['SG300'] = nouns_split['noun'].apply(
    lambda x: cbow200.wv[x]), nouns_split['noun'].apply(lambda x: cbow300.wv[x]), nouns_split['noun'].apply(
    lambda x: sg200.wv[x]), nouns_split['noun'].apply(lambda x: sg300.wv[x])
nouns_split

nouns_split['CBOW200'], nouns_split['CBOW300'], nouns_split['SG200'], nouns_split['SG300'] = nouns_split[
                                                                                                 'CBOW200'].apply(
    lambda x: ' '.join([str(number) for number in x])), nouns_split['CBOW300'].apply(
    lambda x: ' '.join([str(number) for number in x])), nouns_split['SG200'].apply(
    lambda x: ' '.join([str(number) for number in x])), nouns_split['SG300'].apply(
    lambda x: ' '.join([str(number) for number in x]))
nouns_split

nouns_split.to_excel('../data/word2vec_min_count_100.xlsx', index=False)

len(cbow200.wv.index_to_key)
