==> ZhouZhiwei/kansei_bj_sh_hotel/code/7_tencent_wv.ipynb <==
"""
腾讯预训练词向量
"""
import pandas as pd
from gensim.models import KeyedVectors

data = pd.read_excel('../data/cluster_cleaned.xlsx')
data

tencent_wv = KeyedVectors.load_word2vec_format(
    r'E:\OneDrive - tgblx\Python\wv\tencent-ailab-embedding-zh-d200-v0.2.0\tencent-ailab-embedding-zh-d200-v0.2.0.txt',
    binary=False)

data = data[data['noun'].apply(lambda x: x in tencent_wv.index_to_key)]
data.shape

data['tencent200'] = data['noun'].apply(lambda x: tencent_wv[x])
data

data['tencent200'] = data['tencent200'].apply(lambda x: ' '.join([str(number) for number in x]))
data

data.to_excel('../data/mywv_wwm_tencent_count100.xlsx',index=False)
