==> ZhouZhiwei/kansei_bj_sh_hotel/code/8_RoBERTa_wwm.ipynb <==
"""
生成BERT词向量，模型为 hfl/chinese-roberta-wwm-ext
"""
from transformers import BertTokenizer, BertModel
import pandas as pd

count5 = pd.read_excel('../data/word2vec_min_count_5.xlsx')
count5

bert_tokenizer = BertTokenizer.from_pretrained("hfl/chinese-roberta-wwm-ext")
bert_model = BertModel.from_pretrained("hfl/chinese-roberta-wwm-ext")


def get_word_vector(text, tokenizer, model):
    encoded_input = tokenizer(text, return_tensors='pt')
    output = model(**encoded_input)
    # 取出tensor中的值
    return ' '.join([str(number) for number in output.last_hidden_state.tolist()[0][0]])

count5['roberta_wwm'] = count5['noun'].apply(
    lambda x: get_word_vector(text=x, tokenizer=bert_tokenizer, model=bert_model))
count5

count5.to_excel('../data/word2vec_bert_count5.xlsx', index=False)

count1 = pd.read_excel('../data/word2vec_min_count_1.xlsx')
count1['roberta_wwm'] = count1['noun'].apply(
    lambda x: get_word_vector(text=x, tokenizer=bert_tokenizer, model=bert_model))
count1.to_excel('../data/word2vec_bert_count1.xlsx', index=False)
count1

count1

count1 = count1[~count1['noun'].isna()]
count1.shape

count1['roberta_wwm'] = count1['noun'].apply(
    lambda x: get_word_vector(text=x, tokenizer=bert_tokenizer, model=bert_model))
count1.to_excel('../data/word2vec_bert_count1.xlsx', index=False)
count1
