==> ZhouZhiwei/kansei_bj_sh_hotel/code/10_sentence_BERT.ipynb <==
'''
sentence-transformer embedding
model:paraphrase-multilingual-mpnet-base-v2
'''
from sentence_transformers import SentenceTransformer
import pandas as pd

# 本地导入模型指定模型路径
model = SentenceTransformer('E:\OneDrive - tgblx\Python\model\paraphrase-multilingual-mpnet-base-v2')

data = pd.read_excel('../data/mywv_wwm_tencent_count100.xlsx')
data

data['sentence_transformer'] = data['noun'].apply(lambda x: model.encode(x))
data

data['sentence_transformer'] = data['sentence_transformer'].apply(lambda x: ' '.join([str(number) for number in x]))
data

data.to_excel('../data/mywv_wwm_tencent_sentence_count100.xlsx', index=False)
