==> ZhouZhiwei/kansei_bj_sh_hotel/code/11_clustering.ipynb <==
"""
名词聚类
"""
import pandas as pd
from sklearn import metrics
from sklearn.cluster import KMeans, AgglomerativeClustering

data = pd.read_excel('../data/word2vec_bert_count100.xlsx')
data

cbow200 = data['CBOW200'].apply(lambda x: [float(number) for number in x.split(' ')]).tolist()
cbow300 = data['CBOW300'].apply(lambda x: [float(number) for number in x.split(' ')]).tolist()
sg200 = data['SG200'].apply(lambda x: [float(number) for number in x.split(' ')]).tolist()
sg300 = data['SG300'].apply(lambda x: [float(number) for number in x.split(' ')]).tolist()
bert = data['roberta_wwm'].apply(lambda x: [float(number) for number in x.split(' ')]).tolist()

# CBOW200 KMeans
cbow200_kmeans = KMeans(n_clusters=17, random_state=42).fit(cbow200)
cbow200_kmeans_labels = cbow200_kmeans.labels_
cbow200_kmeans_dvi = metrics.davies_bouldin_score(cbow200, cbow200_kmeans_labels)
cbow200_kmeans_sc = metrics.silhouette_score(cbow200, cbow200_kmeans_labels)
cbow200_kmeans_chi = metrics.calinski_harabasz_score(cbow200, cbow200_kmeans_labels)

# CBOW300 KMeans
cbow300_kmeans = KMeans(n_clusters=17, random_state=42).fit(cbow300)
cbow300_kmeans_labels = cbow300_kmeans.labels_
cbow300_kmeans_dvi = metrics.davies_bouldin_score(cbow300, cbow300_kmeans_labels)
cbow300_kmeans_sc = metrics.silhouette_score(cbow300, cbow300_kmeans_labels)
cbow300_kmeans_chi = metrics.calinski_harabasz_score(cbow300, cbow300_kmeans_labels)

# SG200 KMeans
sg200_kmeans = KMeans(n_clusters=17, random_state=42).fit(sg200)
sg200_kmeans_labels = sg200_kmeans.labels_
sg200_kmeans_dvi = metrics.davies_bouldin_score(sg200, sg200_kmeans_labels)
sg200_kmeans_sc = metrics.silhouette_score(sg200, sg200_kmeans_labels)
sg200_kmeans_chi = metrics.calinski_harabasz_score(sg200, sg200_kmeans_labels)

# SG300 KMeans
sg300_kmeans = KMeans(n_clusters=17, random_state=42).fit(sg300)
sg300_kmeans_labels = sg300_kmeans.labels_
sg300_kmeans_dvi = metrics.davies_bouldin_score(sg300, sg300_kmeans_labels)
sg300_kmeans_sc = metrics.silhouette_score(sg300, sg300_kmeans_labels)
sg300_kmeans_chi = metrics.calinski_harabasz_score(sg300, sg300_kmeans_labels)

# CBOW200 Hierarchy
cbow200_h = AgglomerativeClustering(n_clusters=17).fit(cbow200)
cbow200_h_labels = cbow200_h.labels_
cbow200_h_dvi = metrics.davies_bouldin_score(cbow200, cbow200_h_labels)
cbow200_h_sc = metrics.silhouette_score(cbow200, cbow200_h_labels)
cbow200_h_chi = metrics.calinski_harabasz_score(cbow200, cbow200_h_labels)

# CBOW300 Hierarchy
cbow300_h = AgglomerativeClustering(n_clusters=17).fit(cbow300)
cbow300_h_labels = cbow300_h.labels_
cbow300_h_dvi = metrics.davies_bouldin_score(cbow300, cbow300_h_labels)
cbow300_h_sc = metrics.silhouette_score(cbow300, cbow300_h_labels)
cbow300_h_chi = metrics.calinski_harabasz_score(cbow300, cbow300_h_labels)

# SG200 Hierarchy
sg200_h = AgglomerativeClustering(n_clusters=17).fit(sg200)
sg200_h_labels = sg200_h.labels_
sg200_h_dvi = metrics.davies_bouldin_score(sg200, sg200_h_labels)
sg200_h_sc = metrics.silhouette_score(sg200, sg200_h_labels)
sg200_h_chi = metrics.calinski_harabasz_score(sg200, sg200_h_labels)

# SG300 Hierarchy
sg300_h = AgglomerativeClustering(n_clusters=17).fit(sg300)
sg300_h_labels = sg300_h.labels_
sg300_h_dvi = metrics.davies_bouldin_score(sg300, sg300_h_labels)
sg300_h_sc = metrics.silhouette_score(sg300, sg300_h_labels)
sg300_h_chi = metrics.calinski_harabasz_score(sg300, sg300_h_labels)

data['CBOW200_kmeans'], data['CBOW300_kmeans'], data['SG200_kmeans'], data['SG300_kmeans'], data['CBOW200_h'], data[
    'CBOW300_h'], data['SG200_h'], data[
    'SG300_h'] = cbow200_kmeans_labels, cbow300_kmeans_labels, sg200_kmeans_labels, sg300_kmeans_labels, cbow200_h_labels, cbow300_h_labels, sg200_h_labels, sg300_h_labels
data

bert_kmeans = KMeans(n_clusters=17, random_state=42).fit(bert)
bert_kmeans_labels = bert_kmeans.labels_
bert_kmeans_dvi = metrics.davies_bouldin_score(bert, bert_kmeans_labels)
bert_kmeans_sc = metrics.silhouette_score(bert, bert_kmeans_labels)
bert_kmeans_chi = metrics.calinski_harabasz_score(bert, bert_kmeans_labels)

bert_h = AgglomerativeClustering(n_clusters=17).fit(bert)
bert_h_labels = bert_h.labels_
bert_h_dvi = metrics.davies_bouldin_score(bert, bert_h_labels)
bert_h_sc = metrics.silhouette_score(bert, bert_h_labels)
bert_h_chi = metrics.calinski_harabasz_score(bert, bert_h_labels)

data['bert_kmeans'], data['bert_h'] = bert_kmeans_labels, bert_h_labels
data

data.to_excel('../data/cluster.xlsx', index=False)

score = {
    'cbow200_k': [cbow200_kmeans_dvi, cbow200_kmeans_sc, cbow200_kmeans_chi],
    'cbow300_k': [cbow300_kmeans_dvi, cbow300_kmeans_sc, cbow300_kmeans_chi],
    'sg200_k': [sg200_kmeans_dvi, sg200_kmeans_sc, sg200_kmeans_chi],
    'sg300_k': [sg300_kmeans_dvi, sg300_kmeans_sc, sg300_kmeans_chi],
    'cbow200_h': [cbow200_h_dvi, cbow200_h_sc, cbow200_h_chi],
    'cbow300_h': [cbow300_h_dvi, cbow300_h_sc, cbow300_h_chi],
    'sg200_h': [sg200_h_dvi, sg200_h_sc, sg200_h_chi],
    'sg300_h': [sg300_h_dvi, sg300_h_sc, sg300_h_chi],
    'bert_k': [bert_kmeans_dvi, bert_kmeans_sc, bert_kmeans_chi],
    'bert_h': [bert_h_dvi, bert_h_sc, bert_h_chi]
}
score

metric_df = pd.DataFrame(score)
metric_df

metric_df.to_excel('../data/metric.xlsx', index=False)
