==> ZhouZhiwei/kansei_bj_sh_hotel/code/12_reduction_clustering.ipynb <==
"""
词向量降维聚类
"""
import umap
import hdbscan
import pandas as pd
from sklearn import metrics
from sklearn.cluster import KMeans, AgglomerativeClustering

data = pd.read_excel('../data/mywv_wwm_tencent_sentence_count100.xlsx')
data

reducer = umap.UMAP(n_neighbors=15, n_components=5, metric='cosine', random_state=42)
metric = pd.DataFrame()
for name, column in data.iteritems():
    if name == 'noun':
        metric['metric'] = ['dvi', 'sc', 'chs', 'class_number']
        continue

    # 类型转换
    embedding_list = column.apply(lambda x: [float(number) for number in x.split(' ')]).tolist()

    # 降维
    reduced = reducer.fit_transform(embedding_list)

    # 聚类 k:kmeans, h:AgglomerativeClustering, hd:hdbscan
    reduced_k = KMeans(n_clusters=35, random_state=42).fit(reduced)
    original_k = KMeans(n_clusters=35, random_state=42).fit(embedding_list)
    reduced_h = AgglomerativeClustering(n_clusters=35).fit(reduced)
    original_h = AgglomerativeClustering(n_clusters=35).fit(embedding_list)
    reduced_hd = hdbscan.HDBSCAN(min_cluster_size=10, min_samples=1).fit(reduced)
    original_hd = hdbscan.HDBSCAN(min_cluster_size=4, min_samples=1).fit(embedding_list)

    # metric
    # hdbscan去掉噪音点计算metric
    non_noise_re, non_noise_labels_re = [], []
    for sample, label in zip(reduced, reduced_hd.labels_):
        if label != -1:
            non_noise_re.append(sample)
            non_noise_labels_re.append(label)

    non_noise, non_noise_labels = [], []
    for sample, label in zip(embedding_list, original_hd.labels_):
        if label != -1:
            non_noise.append(sample)
            non_noise_labels.append(label)

    dvi_re_k = metrics.davies_bouldin_score(reduced, reduced_k.labels_)
    sc_re_k = metrics.silhouette_score(reduced, reduced_k.labels_)
    chs_re_k = metrics.calinski_harabasz_score(reduced, reduced_k.labels_)

    dvi_k = metrics.davies_bouldin_score(embedding_list, original_k.labels_)
    sc_k = metrics.silhouette_score(embedding_list, original_k.labels_)
    chs_k = metrics.calinski_harabasz_score(embedding_list, original_k.labels_)

    dvi_re_h = metrics.davies_bouldin_score(reduced, reduced_h.labels_)
    sc_re_h = metrics.silhouette_score(reduced, reduced_h.labels_)
    chs_re_h = metrics.calinski_harabasz_score(reduced, reduced_h.labels_)

    dvi_h = metrics.davies_bouldin_score(embedding_list, original_h.labels_)
    sc_h = metrics.silhouette_score(embedding_list, original_h.labels_)
    chs_h = metrics.calinski_harabasz_score(embedding_list, original_h.labels_)

    dvi_re_hd = metrics.davies_bouldin_score(non_noise_re, non_noise_labels_re)
    sc_re_hd = metrics.silhouette_score(non_noise_re, non_noise_labels_re)
    chs_re_hd = metrics.calinski_harabasz_score(non_noise_re, non_noise_labels_re)

    dvi_hd = metrics.davies_bouldin_score(non_noise, non_noise_labels)
    sc_hd = metrics.silhouette_score(non_noise, non_noise_labels)
    chs_hd = metrics.calinski_harabasz_score(non_noise, non_noise_labels)

    # 保存聚类结果
    data[name + '_re_k'] = reduced_k.labels_
    data[name + '_k'] = original_k.labels_
    data[name + '_re_h'] = reduced_h.labels_
    data[name + '_h'] = original_h.labels_
    data[name + '_re_hd'] = reduced_hd.labels_
    data[name + '_hd'] = original_hd.labels_

    metric[name + '_re_k'] = [dvi_re_k, sc_re_k, chs_re_k, reduced_k.labels_.max() +1]
    metric[name + '_k'] = [dvi_k, sc_k, chs_k, original_k.labels_.max() + 1]
    metric[name + '_re_h'] = [dvi_re_h, sc_re_h, chs_re_h, reduced_h.labels_.max() + 1]
    metric[name + '_h'] = [dvi_h, sc_h, chs_h, original_h.labels_.max() + 1]
    metric[name + '_re_hd'] = [dvi_re_hd, sc_re_hd, chs_re_hd, reduced_hd.labels_.max() + 1]
    metric[name + '_hd'] = [dvi_hd, sc_hd, chs_hd, original_hd.labels_.max() + 1]

data

metric

data.to_excel('../data/cluster_reduced_sentence.xlsx', index=False)
metric.to_excel('../data/metric_reduced_sentence.xlsx', index=False)
