==> ZhouZhiwei/kansei_bj_sh_hotel/code/13_cluster_export.ipynb <==
"""
聚类结果导出
"""
import pandas as pd

data = pd.read_excel('../data/cluster_reduced_sentence.xlsx')
data

cluster_cbow200_hdbscan = data[['noun', 'CBOW200_re_hd']]
cluster_sorted = cluster_cbow200_hdbscan.sort_values(by='CBOW200_re_hd')
group = cluster_cbow200_hdbscan.groupby('CBOW200_re_hd')
df_hd = group.apply(lambda x: '，'.join(x['noun'].tolist()))
df_hd.to_excel('../data/clusters.xlsx')

cluster_cbow200_kmeans = data[['noun', 'CBOW200_re_k']]
group = cluster_cbow200_kmeans.groupby('CBOW200_re_k')
df_k = group.apply(lambda x: '，'.join(x['noun'].tolist()))
df_k.to_excel('../data/clusters_cbow200_kmeans_original.xlsx')

cluster_cbow200_hierarchy = data[['noun', 'CBOW200_re_h']]
group = cluster_cbow200_hierarchy.groupby('CBOW200_re_h')
df_h = group.apply(lambda x: '，'.join(x['noun'].tolist()))
df_h.to_excel('../data/clusters_cbow200_hierarchy.xlsx')
