==> ZhouZhiwei/kansei_bj_sh_hotel/code/20_sentence_topic_match_similarity.ipynb <==
'''
句子主题匹配，词典匹配不到使用词向量相似度匹配。句子向量为词向量平均。
'''
import gensim.models
import pandas as pd

# 读词向量CBOW200
model = gensim.models.Word2Vec.load('../model/gensim-model-43gruuh9-CBOW200-mincount100')
word_vector = model.wv

# 读未匹配文本
data = pd.read_csv('../data/sentence_match_result_without_match_2.csv', sep='\t')

# 读取停用词
with open('../data/baidu_stopwords.txt', 'r', encoding='utf-8') as f:
    stop_list = f.read()
    # 返回list类型数据
    stop_list = stop_list.split('\n')

# 读取主题词
topic = pd.read_excel('../data/4_clusters_cbow200_kmeans-words_fixed-Wu.xlsx')
topic['Phrases'] = topic['Phrases'].apply(lambda x: x.split('，'))
all_word = topic['Phrases'].tolist()

# 构造词典：{word->topic}
word_dict = {}
for i, topic_words in enumerate(all_word):
    for topic_word in topic_words:
        word_dict[topic_word] = i

topic_word_df = pd.DataFrame()
topic_word_df['word'] = word_dict.keys()
topic_word_df['topic'] = word_dict.values()

from LAC import LAC

# 初始化分词模型
lac = LAC(mode='seg')


# 处理文本：分词，去停用词
def text_process(text):
    word_list = lac.run(text)
    word_list = list(filter(lambda x: x not in stop_list, word_list))
    word_list = list(filter(lambda x: x in word_vector, word_list))
    return word_list


# 生成句向量：对句中词向量取平均
def sentence_vector(words):
    word_num = len(words)
    vector_sum = 0
    for word in words:
        vector_sum += word_vector[word]
    return vector_sum / word_num

from tqdm import tqdm

tqdm.pandas()
data['word'] = data['sentence'].progress_apply(text_process)
data.shape

data = data[data['word'].progress_apply(lambda x: len(x) != 0)]
data.shape

data['embedding'] = data['word'].progress_apply(sentence_vector)
data

topic_word_df['embedding'] = topic_word_df['word'].apply(lambda x: word_vector[x])
topic_word_df

from sklearn.metrics.pairwise import cosine_similarity
import numpy as np

topic_word_embedding = topic_word_df['embedding'].tolist()


def match(embedding):
    embedding = embedding.reshape(1, -1)
    cos_sim = cosine_similarity(embedding, topic_word_embedding)
    max_index = np.argmax(cos_sim)
    result = topic_word_df.iloc[max_index]['topic']
    return result

data['topic'] = data['embedding'].progress_apply(match)
data

data = data.drop(labels=['word', 'embedding'], axis=1)
data.to_csv('../data/sentence_match_result_similarity_2.csv', sep='\t', index=False)

match_dict = pd.read_csv('../data/sentence_match_result_dict_2.csv', sep='\t')
match_dict.head()

match_dict['match'] = 'dict'
data['match'] = 'similarity'
data_all = pd.concat([match_dict, data])
data_all.shape

data_all.to_csv('../data/sentence_match_result_2.csv', sep='\t', index=False)

# 计算performance
senti = data_all[['sentiment', 'topic']]
groups = senti.groupby(by='topic')

mean = groups.mean()
mean.columns = ['performance']
mean['performance_std'] = (mean['performance'] - mean['performance'].min()) / (
            mean['performance'].max() - mean['performance'].min())
mean

mean.to_csv('../data/performance.csv', sep='\t')
